[CPSRV-17563] Emit sql mig content on console if dry-run is enabled

+++ src/Phinx/Db/Adapter/MysqlAdapter.php
@@ -1120,6 +1120,12 @@ class MysqlAdapter extends PdoAdapter implements AdapterInterface
      */
     public function runUsingBinary($sql)
     {
+        if ($this->isDryRunEnabled()) {
+            $this->getOutput()->writeln($sql);
+
+            return;
+        }
+
         $tempFile = tempnam(sys_get_temp_dir(), 'phinx');
         file_put_contents($tempFile, $sql);
